new class for code related to SQLite db cache
