+++ Database/TestDatabaseCache.php
@@ -1,0 +1,106 @@
+<?php
+
+/*
+ * This file is part of the Liip/FunctionalTestBundle
+ *
+ * (c) Lukas Kahwe Smith <smith@pooteeweet.org>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Liip\FunctionalTestBundle\Database;
+
+use Symfony\Component\DependencyInjection\Container;
+
+class TestDatabaseCache
+{
+    private $container;
+
+    public function __construct(Container $container)
+    {
+        $this->container = $container;
+    }
+
+    /**
+     * This function finds the time when the data blocks of a class definition
+     * file were being written to, that is, the time when the content of the
+     * file was changed.
+     *
+     * @param string $class
+     *            The fully qualified class name of the fixture class to
+     *            check modification date on.
+     *
+     * @return \DateTime|null
+     */
+    protected function getClassLastModified($class)
+    {
+        $lastModifiedDateTime = null;
+
+        $reflClass = new \ReflectionClass($class);
+        $classFileName = $reflClass->getFileName();
+
+        if (file_exists($classFileName)) {
+            $lastModifiedDateTime = new \DateTime();
+            $lastModifiedDateTime->setTimestamp(filemtime($classFileName));
+        }
+
+        return $lastModifiedDateTime;
+    }
+
+    /**
+     * Determine if the Fixtures that define a database backup have been
+     * modified since the backup was made.
+     *
+     * @param array $classNames
+     *            The fixture classnames to check
+     * @param string $backup
+     *            The fixture backup SQLite database file path
+     *
+     * @return bool TRUE if the backup was made since the modifications to the
+     *         fixtures; FALSE otherwise
+     */
+    public function isBackupUpToDate(array $classNames, $backup)
+    {
+        if(!file_exists($backup) || !file_exists($backup.'.ser')) {
+            return false;
+        }
+
+        $backupLastModifiedDateTime = new \DateTime();
+        $backupLastModifiedDateTime->setTimestamp(filemtime($backup));
+
+        foreach ($classNames as &$className) {
+            $fixtureLastModifiedDateTime = $this->getClassLastModified($className);
+            if ($backupLastModifiedDateTime < $fixtureLastModifiedDateTime) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    /**
+     *
+     * @param array $metadatas
+     * @param array $classNames
+     * @return string full path to the database cache file
+     */
+    public function buildCacheFilePath(array $metadatas, array $classNames)
+    {
+        return $this->container->getParameter('kernel.cache_dir') . '/test_' . md5(serialize($metadatas) . serialize($classNames)) . '.db';
+    }
+
+    /**
+     * Is caching of SQLite Database instances enabled?
+     *
+     * Caching can be enabled in app/config/config_test.yml:
+     * ``` yaml
+     * liip_functional_test:
+     *     cache_sqlite_db: true
+     * ```
+     */
+    public function isCacheEnabled()
+    {
+        return $this->container->getParameter('liip_functional_test.cache_sqlite_db');
+    }
+}
