+++ Test/WebTestCase.php
@@ -35,6 +35,8 @@ use Doctrine\DBAL\Driver\PDOSqlite\Driver as SqliteDriver;
 
 use Doctrine\ORM\Tools\SchemaTool;
 
+use Liip\FunctionalTestBundle\Database\TestDatabaseCache;
+
 /**
  * @author Lea Haensenberger
  * @author Lukas Kahwe Smith <smith@pooteeweet.org>
@@ -159,56 +161,6 @@ abstract class WebTestCase extends BaseWebTestCase
     }
 
     /**
-     * This function finds the time when the data blocks of a class definition
-     * file were being written to, that is, the time when the content of the
-     * file was changed.
-     *
-     * @param string $class The fully qualified class name of the fixture class to
-     * check modification date on.
-     *
-     * @return \DateTime|null
-     */
-    protected function getClassLastModified($class)
-    {
-        $lastModifiedDateTime = null;
-
-        $reflClass = new \ReflectionClass($class);
-        $classFileName = $reflClass->getFileName();
-
-        if (file_exists($classFileName)) {
-            $lastModifiedDateTime = new \DateTime();
-            $lastModifiedDateTime->setTimestamp(filemtime($classFileName));
-        }
-
-        return $lastModifiedDateTime;
-    }
-
-    /**
-     * Determine if the Fixtures that define a database backup have been
-     * modified since the backup was made.
-     *
-     * @param array  $classNames The fixture classnames to check
-     * @param string $backup     The fixture backup SQLite database file path
-     *
-     * @return bool TRUE if the backup was made since the modifications to the
-     * fixtures; FALSE otherwise
-     */
-    protected function isBackupUpToDate(array $classNames, $backup)
-    {
-        $backupLastModifiedDateTime = new \DateTime();
-        $backupLastModifiedDateTime->setTimestamp(filemtime($backup));
-
-        foreach ($classNames as &$className) {
-            $fixtureLastModifiedDateTime = $this->getClassLastModified($className);
-            if ($backupLastModifiedDateTime < $fixtureLastModifiedDateTime) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    /**
      * @param string $registryName, e.g. 'doctrine'
      * @return ManagerRegistry
      */
@@ -283,9 +235,10 @@ abstract class WebTestCase extends BaseWebTestCase
                 }
                 $metadatas = self::$cachedMetadatas[$omName];
 
-                if ($container->getParameter('liip_functional_test.cache_sqlite_db')) {
-                    $backup = $container->getParameter('kernel.cache_dir') . '/test_' . md5(serialize($metadatas) . serialize($classNames)) . '.db';
-                    if (file_exists($backup) && file_exists($backup.'.ser') && $this->isBackupUpToDate($classNames, $backup)) {
+                $dbCache = new TestDatabaseCache($container);
+                if ($dbCache->isCacheEnabled()) {
+                    $backup = $dbCache->buildCacheFilePath($metadatas, $classNames);
+                    if ($dbCache->isBackupUpToDate($classNames, $backup)) {
                         $om->flush();
                         $om->clear();
